+++ NEWS.md
@@ -1,6 +1,7 @@
 - v0.7.0
   - Implement `mod2pi` for degrees, cleanup display of degree units.
   - Tweak implementation of `Gain` types for usability.
+  - Implement `zero` and `one` for `Level` and `Gain`.
 - v0.6.1
   - Permit symbols that are bound to `Number`s to be used in `u_str` macro, such that
     π and other non-literal numbers can be used.